+++ packages/react/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @token-icons/react
 
+## 2.3.1
+
+### Patch Changes
+
+-
+- Updated dependencies
+  - @token-icons/core@2.3.1
+
+## 2.3.0
+
+### Minor Changes
+
+- new network icons: x-layer, etherlink
+
+### Patch Changes
+
+- Updated dependencies
+  - @token-icons/core@2.3.0
+
 ## 2.2.2
 
 ### Patch Changes